rename folder: ErrorBanner -> ErrorAlert
