+++ src/aics-image-viewer/components/ErrorAlert/index.tsx
